+++ patterns/header/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.0.0-alpha.4
+
+### Patch Changes
+
+- Updated dependencies [88e2daee]
+  - react-magma-dom@2.4.0-alpha.4
+
 ## 1.0.0-alpha.3
 
 ### Patch Changes